requirements.txt
Django==2.1.4
django-extensions==2.2.1
django-filter==2.2.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.2
six==1.12.0
